LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-constructor-expected.txt

Testing the IOSInlineMediaControls constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.element.localName is "div"
PASS mediaControls.element.className is "media-controls ios inline"
PASS mediaControls.layoutTraits is LayoutTraits.iOS

Controls bar
PASS mediaControls.controlsBar.children.length is 2
PASS mediaControls.controlsBar.children[0] instanceof ButtonsContainer is true
PASS mediaControls.controlsBar.children[1] instanceof ButtonsContainer is true

Left container
PASS mediaControls.controlsBar.children[0].buttons is [mediaControls.playPauseButton, mediaControls.skipBackButton]
PASS mediaControls.controlsBar.children[0].element.className is "buttons-container left"
PASS mediaControls.controlsBar.children[0].leftMargin is 24
PASS mediaControls.controlsBar.children[0].rightMargin is 24
PASS mediaControls.controlsBar.children[0].buttonMargin is 24

Right container
PASS mediaControls.controlsBar.children[1].buttons is [mediaControls.airplayButton, mediaControls.pipButton, mediaControls.fullscreenButton]
PASS mediaControls.controlsBar.children[1].element.className is "buttons-container right"
PASS mediaControls.controlsBar.children[1].leftMargin is 24
PASS mediaControls.controlsBar.children[1].rightMargin is 24
PASS mediaControls.controlsBar.children[1].buttonMargin is 24

PASS successfullyParsed is true

TEST COMPLETE
